Grafana + Loki + Alloy on Local env
Enables Grafana, Loki and Alloy on the local Kind environment with required config, so that these components work together out-of-the-box.

A simple Alloy pipeline using the `loki` component ensures that Alloy scrapes all pod logs. These logs are send to Loki. Grafana is deployed in the simplest way possible, with default login credentials and a Loki datasource included.