feature: Grafana + Mimir + Alloy on Local env
This MR enables Grafana, Mimir and Alloy on the local Kind environment with required config, so that these components work together out-of-the-box. As Mimir is very resource intensive, a forked Helm chart has been used which makes it possible to run Mimir in monolithic mode. For details, see the added README in the local Mimir overlay.

A simple Alloy metrics pipeline ensures that Alloy scrapes the most important cluster metrics. These metrics are sent to Mimir. Grafana is deployed in the simplest way possible, with default login credentials and a Mimir (prometheus) datasource included. Dashboards displaying K8S metrics have already been added in the past.